In the process, also re-instates PPC and OS 10.4 support.  The requirements for compiling this are now noted in the XCConfig file.
Also bumped the version number.

git-svn-id: https://soundflower.googlecode.com/svn/trunk@84 0051ed44-974d-0410-95fd-975f363cc5de

+++ Source/SoundflowerEngine.cpp
@@ -305,7 +305,7 @@ bool SoundflowerEngine::createAudioStreams(IOAudioSampleRate *initialSampleRate)
             outputBuffer = (void *)IOMalloc(outputBufferSize);
             
             if (!outputBuffer) {
-                IOLog("Soundflower: Error allocating output buffer - %lu bytes.\n", outputBufferSize);
+                IOLog("Soundflower: Error allocating output buffer - %lu bytes.\n", (unsigned long)outputBufferSize);
                 goto Error;
             }
 
@@ -314,7 +314,7 @@ bool SoundflowerEngine::createAudioStreams(IOAudioSampleRate *initialSampleRate)
             thruBufferSize = outputBufferSize;
             thruBuffer = (float *)IOMalloc(thruBufferSize);
             if (!thruBuffer) {
-                IOLog("Soundflower: Error allocating thru buffer - %lu bytes.\n", thruBufferSize);
+                IOLog("Soundflower: Error allocating thru buffer - %lu bytes.\n", (unsigned long)thruBufferSize);
                 goto Error;
             }
             memset ((UInt8 *)thruBuffer, 0, thruBufferSize);
@@ -325,7 +325,7 @@ bool SoundflowerEngine::createAudioStreams(IOAudioSampleRate *initialSampleRate)
             if (separateInputBuffers) {
                 inputBuffer = (void *)IOMalloc(inputBufferSize);
                 if (!inputBuffer) {
-                    IOLog("Soundflower: Error allocating input buffer - %lu bytes.\n", inputBufferSize);
+                    IOLog("Soundflower: Error allocating input buffer - %lu bytes.\n", (unsigned long )inputBufferSize);
                     goto Error;
                 }
             } else {
